+++ src/components/SingleDatePicker.vue
@@ -9,7 +9,6 @@
 
 <script>
 import Calendar from './Calendar';
-import defaults from '../utils/defaults';
 import { singleHasValue, singleValuesAreEqual } from '../utils/pickerProfiles';
 
 export default {
@@ -18,7 +17,6 @@ export default {
   },
   props: {
     value: { type: Date, default: null },
-    selectColor: { type: String, default: () => defaults.datePickerSelectColor },
     selectAttribute: Object,
     disabledAttribute: Object,
     attributes: Array,
@@ -27,7 +25,6 @@ export default {
     selectAttribute_() {
       if (!singleHasValue(this.value)) return null;
       return {
-        ...defaults.datePickerSelectAttribute(this.selectColor),
         ...this.selectAttribute,
         dates: [this.value],
       };